2015 ACM-ICPC, Ferry crossing problem
Ferries crossing the Strait of Gibraltar from Morocco to Spain must
carefully navigate to avoid the
heavy ship traffic along the strait. Write a program to help ferry
captains find the largest gaps
in strait traffic for a safe crossing. Your program will use a simple
model as follows. The strait
has several parallel shipping lanes in east- west direction. Ships run
with the same constant speed
either eastbound or westbound. All ships in the same lane run in the
same direction. Satellite data
provides the positions of the ships in each lane. The ships may have
different lengths. Ships do not
change lanes and do not change speed for the crossing ferry. The ferry
waits for an appropriate time
when there is an adequate gap in the ship traffic. It then crosses the
strait heading northbound
along a north-south line at a constant speed. From the moment a ferry
enters a lane until the
moment it leaves the lane, no ship in that lane may touch the crossing
line. Ferries are so small
you can neglect their size. Figure I.1 illustrates the lanes and ships
for Sample Input 1. Your task
Problem I Ship Traffic
is to find the largest time interval within which the ferry can safely
cross the strait.